+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* Fix `Command::is_enabled` and `has_handlers` sometimes entering wrong state.
+
 * Change `zng::task::block_on` to never run tasks in the app context.
     - Now logs a warning and removes the app context.
     - Fixes deadlock when blocking on `zng::task::deadline`.